pyproject.toml
[project]
name = "geomscreen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ase>=3.23",
    "dplutils>=0.8.2",
    "fairchem-core>=2.14.0",
    "pandas>=3",
]

[build-system]
requires = ["uv_build>=0.9.29,<0.10.0"]
build-backend = "uv_build"
